feat: add device login flow
